requirements.txt
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
idna==3.10
msal==1.32.3
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.4
urllib3==2.5.0
streamlit==1.46.1
google-genai==1.19.0
pytest==8.4.1